Make illegal slot instructions save the address of the branch
With no patches, J2 would save on the stack the address of the branch target
when an illegal slot instruction was encountered. This is not very useful
because it makes it impossible to resume the branch after fixing up the
illegal instruction somehow. With the previous patches, J2 would save the
address of the delay slot + 4 which is weird but workable. Make the microcode
for illegal slot instructions subtract 6 so that the saved address is now the
address of the branch.